requirements.txt
# Python==3.10
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
imageio==2.37.0
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.7
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
tifffile==2025.5.10
tzdata==2025.2
